move test files, add more unit tests

+++ utils.js
@@ -12,10 +12,12 @@ require('common-middleware', 'middleware');
 require('common-questions', 'questions');
 require('empty-dir');
 require('extend-shallow', 'extend');
+require('fs-exists-sync', 'exists');
 require('is-valid-app', 'isValid');
 require = fn;
 
 utils.isEmpty = function(filepath) {
+  if (!utils.exists(filepath)) return true;
   return utils.emptyDir.sync(filepath, function(fp) {
     return !/(\.DS_Store|Thumbs\.db)/.test(fp);
   });
